pyproject.toml
[project]
name = "feedback-to-me"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.2.1",
    "google-generativeai>=0.8.4",
    "langchain-google-genai>=2.0.9",
    "langchain>=0.3.17",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
    "python-fasthtml>=0.12.1",
    "stripe>=11.5.0",
    "google-genai>=1.1.0",
    "psycopg2-binary>=2.9.10",
    "fastsql>=2.0.3",
]
